requirements.txt
streamlit
python-pptx
graphviz
